+++ RELEASE_NOTES.md
@@ -7,9 +7,11 @@
 ### Breaking Changes
 
 * Normalized all command line params to camelCase
+* Changed default model to ollama/*
 
 ### New Features
 
+* Added Cloudflare AI Workers for LLM compute
 * Implemented Record and OracleData struct to save events to persisted storage
 * Discord Scraper as pkg
 * Twitter Scraper as pkg